% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_rect.R
\name{exp_rect}
\alias{exp_rect}
\title{Expected number of rectangles}
\usage{
exp_rect(x_coord, y_coord, mindist, maxdist, pointjitter)
}
\arguments{
\item{x_coord}{Vector of x-coordinates.}

\item{y_coord}{Vector of y-coordinates.}

\item{mindist}{Decimal, minimum distance between postholes.}

\item{maxdist}{Decimal, maximum distance between postholes.}

\item{pointjitter}{Decimal, defining the jittering of points.}
}
\value{
A string explaining the number of expected rectangles.
}
\description{
\code{exp_rect} allows the user to determine the expected
number of rectangles according to the formula by Fletcher and Lock 1984. This
is important because time of computation rises exponentially with the number
of rectangles. With more than 4000-5000 expected rectangles, computation 
will take unreasonable long (or more likely trigger a memory error).

In comparison to the original formula which had only "0.5" in the bottom of
the formula leading to "NRE", we found that this tends to doubling the true
expected number. Therefore, we changed that this term to "0.25" which produces
numbers in accordance with simulated runs.
}
\examples{
library(spatstat)
sim <- sim_rect(lower = 0, upper = 100, housenumber = 0,  maxrows = 5,
mindist = 2, maxdist = 4, pointjitter = 0, preserve = 100,
rp = 15000, outputmethod = "spatstat")

# exporting the coordinates
sim_coords <- spatstat::coords(sim$points)
x_coord <- sim_coords$x
y_coord <- sim_coords$y

# calculating the number of expected rectangles
exp_rect(x_coord, y_coord, mindist = 1, maxdist = 10, pointjitter = 0.5)

}
