% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rect.R
\name{find_rect}
\alias{find_rect}
\title{Find Rectangles}
\usage{
find_rect(a, b, x, y, z, outputmethod)
}
\arguments{
\item{a}{Vector of x-coordinates.}

\item{b}{Vector of y-coordinates.}

\item{x}{Decimal, minimum distance between postholes.}

\item{y}{Decimal, maximum distance between postholes.}

\item{z}{Decimal, defining the tolerance of points forming rectangles.}

\item{outputmethod}{Character string, defining the format of the output list.
Options include shp (SpatialPointDataframe, SpatialLinesDataframe),
spatstat, spatial}
}
\value{
A list of line and point spatstat objects.
}
\description{
Find Rectangles
}
\examples{
library(spatstat)
library(maptools)
library(sp)
library(data.table)

# setting up the example data set
sim <- sim_rect(lower = 0, upper = 200, housenumber = 20,  maxrows = 5,
mindist = 2, maxdist = 5, pointjitter = 0.3, preserve = 80,
rp = 200, outputmethod = "spatstat")

# exporting the coordinates
psp_coords <- spatstat::coords(sim$points)

# performing the search for rectangles
psp <- find_rect(psp_coords$x,psp_coords$y,2,5,0.3, outputmethod = "spatstat")

# plot the result
plot(psp$points)
plot(psp$lines, add=TRUE)
}
