% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_bound_rect.R
\name{min_bound_rect}
\alias{min_bound_rect}
\title{Minimum bounding rectangle}
\usage{
min_bound_rect(ppp_points, outputmethod)
}
\arguments{
\item{ppp_points}{Points of spatstat::class ppp with marks as factor-level.}

\item{outputmethod}{Character string, defining the format of the output list.
Options include shp (SpatialPointDataframe, SpatialLinesDataframe),
spatstat, spatial}
}
\value{
A list of lines delineating the minimum bounding rectangles of points
marked as belonging together. The format depends on the output method
chosen by the user.
}
\description{
this function finds bounding rectangle. The main body of the 
function derives from this discussion on stackoverflow (dd). Credits go to blutt
}
\examples{
library(spatstat)
library(maptools)
library(sp)
library(data.table)

# setting up the example data set
sim <- sim_rect(lower = 0, upper = 200, housenumber = 20,  maxrows = 5,
mindist = 2, maxdist = 5, pointjitter = 0.3, preserve = 80,
rp = 200, outputmethod = "spatstat")

# exporting the coordinates
psp_coords <- spatstat::coords(sim$points)

# finding the rectangles
psp <- find_rect(psp_coords$x,psp_coords$y, 2, 6, 0.3, outputmethod = "shape")

# preparing the points
ppp_points <- as.ppp(as(psp$lines, "SpatialPointsDataFrame"))

# adding the marks
ppp_points$marks <- as.factor(ppp_points$marks$group_nr)

# performing the analysis
mbr <- min_bound_rect(ppp_points,outputmethod="spatstat")
plot(mbr)

}
