% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAAR.R
\docType{package}
\name{postAAR}
\alias{postAAR}
\alias{postAAR-package}
\title{postAAR: A package for finding and contextualizing rectangles}
\description{
The postAAR package provides following functions:
foo, bar and baz.
}
\section{postAAR functions}{

The foo functions ...
}

